make hyperdrive-math agent
moves all the methods for Agent that require hyeprdrive-math things to
hyperdrive-math/test_utils/agent.rs
