+++ crates/test-utils/Cargo.toml
@@ -26,6 +26,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tokio = { version = "1", features = ["full"] }
 tracing = "0.1"
+getset = "0.1.2"
 
 # Workspace dependencies.
 fixed-point = { path = "../fixed-point" }
